@@ -38,8 +38,8 @@
   <properties>
     <renjin.version>3.5-beta76</renjin.version>
     <renjin.version.for.plugin>3.5-beta76</renjin.version.for.plugin>
-    <maven.compiler.source>1.8</maven.compiler.source>
-    <maven.compiler.target>1.8</maven.compiler.target>
+    <javafx.version>17.0.2</javafx.version>
+    <maven.compiler.release>11</maven.compiler.release>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
   </properties>
